constructor now takes in an double accidentally had it take an integer earlier

+++ src/monopoly/Player.java
@@ -16,7 +16,7 @@ public class Player {
 	// Constructors
 	//
 	
-	public Player(int startMoney) {
+	public Player(double startMoney) {
 		this.money = startMoney;
 		position = 0;
